+++ package.json
@@ -43,8 +43,8 @@
   },
   "lint-staged": {
     "**/*.{ts,tsx}": [
-      "eslint --fix",
-      "prettier --write"
+      "eslint --fix .",
+      "prettier --write --cache ."
     ]
   }
 }
